Billing/requirements.txt
attrs==22.1.0
click==8.1.3
et-xmlfile==1.1.0
exceptiongroup==1.0.1
Flask==2.2.2
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
openpyxl==3.0.10
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
urllib3==1.26.12